requirements.txt
numpy
keras
tensorflow-gpu
progressbar33
pillow
h5py
scikit-learn